+++ arch/arm/mach-tegra/tegra14x_la.c
@@ -109,9 +109,9 @@
 #define T14X_MC_LA_MAX_VALUE		255
 
 #define T14X_MC_RA(r) \
-	((u32)IO_ADDRESS(TEGRA_MC_BASE) + (T14X_MC_##r))
+	(IO_ADDRESS(TEGRA_MC_BASE) + (T14X_MC_##r))
 #define T14X_RA(r) \
-	((u32)IO_ADDRESS(TEGRA_MC_BASE) + (T14X_MC_LA_##r))
+	(IO_ADDRESS(TEGRA_MC_BASE) + (T14X_MC_LA_##r))
 
 #define T14X_LA(f, e, a, r, i, ss, la) \
 { \
